requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Image Processing
pillow>=9.5.0

# Utilities
tqdm>=4.65.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.23.0
